+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 
 * support passing `style` prop from item - #347
 * `selected` is provided to `itemRenderer` - #348
+* simplify logic for calculate dimensions and prevent item width and left properties from being unbounded - (refactoring)
 
 ### 0.17,1